Updated Monster Builder
Updated monster builder to properly allow people to assign a source to
the monster, while moving the primary editor to a JPanel to make it
portable.
