+++ src/builders/monster_builder/MonsterBuilderPane.java
@@ -2,7 +2,6 @@ package builders.monster_builder;
 
 import java.awt.BorderLayout;
 import java.awt.Color;
-import java.awt.Container;
 import java.awt.Dialog;
 import java.awt.Dimension;
 import java.awt.FlowLayout;
@@ -12,19 +11,15 @@ import java.awt.GridBagLayout;
 import java.awt.GridLayout;
 import java.awt.event.MouseEvent;
 import java.awt.event.MouseListener;
-import java.awt.event.WindowEvent;
-import java.awt.event.WindowListener;
-import java.awt.image.BufferedImage;
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 
-import javax.imageio.ImageIO;
 import javax.swing.BorderFactory;
 import javax.swing.JButton;
 import javax.swing.JCheckBox;
+import javax.swing.JComboBox;
 import javax.swing.JDesktopPane;
 import javax.swing.JFrame;
 import javax.swing.JLabel;
@@ -43,17 +38,19 @@ import data.*;
 import data.DataContainer.MapType;
 import data.DataContainer.Proficiency;
 import data.DataContainer.Skills;
+import data.DataContainer.Source;
 import gui.gui_helpers.CompFactory;
 import gui.gui_helpers.DocumentHelper;
 import gui.gui_helpers.MonsterDispPane;
 import gui.gui_helpers.ReminderField;
 import gui.gui_helpers.RichEditor;
+import gui.gui_helpers.CompFactory.ComponentType;
 import gui.gui_helpers.CompFactory.ScrollPolicy;
 import gui.gui_helpers.structures.GuiDirector;
 import gui.gui_helpers.structures.StyleContainer;
 
 @SuppressWarnings("serial")
-public class MonsterBuilder extends JFrame {
+public class MonsterBuilderPane extends JPanel {
 	private DataContainer data;
 
 	private ReminderField monsterNameField, monsterTypeField, acField, initBnsField, hpField, speedField, strField,
@@ -63,6 +60,8 @@ public class MonsterBuilder extends JFrame {
 	private JCheckBox initProfChck, initExpertChck, strSaveP, dexSaveP, conSaveP, intSaveP, wisSaveP, chaSaveP;
 
 	private RichEditor traits, actions, bonusActions, reactions, legendActions;
+	
+	private JComboBox<Source> srcCombo;
 
 	private HashMap<DataContainer.Skills, DataContainer.Proficiency> skills;
 
@@ -80,27 +79,26 @@ public class MonsterBuilder extends JFrame {
 	
 
 	public static void main(String[] args) {
+		DataContainer data = new DataContainer();
+		data.init();
 		SwingUtilities.invokeLater(() -> {
-			MonsterBuilder monstBuild = new MonsterBuilder(new DataContainer());
-			monstBuild.setVisible(true);
+			StyleContainer.SetLookAndFeel();
+			JFrame frm = new JFrame();
+			frm.setDefaultCloseOperation(JFrame.DO_NOTHING_ON_CLOSE);
+			frm.addWindowListener(CompFactory.createSafeExitWindowListener(frm, data));
+			frm.setContentPane(new MonsterBuilderPane(data));
+			frm.pack();
+			frm.setSize(new Dimension(800, 800));
+			frm.setVisible(true);
 		});
 	}
 
-	public MonsterBuilder(DataContainer data) {
+	public MonsterBuilderPane(DataContainer data) {
 		this.data = data;
-		System.out.println(data.getMonsterKeysSorted().size());
 		tagFields = new ArrayList<ReminderField>();
 		skills = new HashMap<Skills, Proficiency>();
-		try {
-			BufferedImage iconImage = ImageIO.read(
-				    getClass().getResource("/"+ StyleContainer.MONSTER_ICON_FILE));
-			this.setIconImage(iconImage);
-		} catch (IOException e) {
-			e.printStackTrace();
-		}
-		
+		ConfigurePane();
 		BuildReplacements();
-		ConfigureFrame(this.getContentPane());
 	}
 	
 	private void BuildReplacements() {
@@ -108,33 +106,9 @@ public class MonsterBuilder extends JFrame {
 		rEditReps.put("<Name>", "");
 	}
 
-	private void ConfigureFrame(Container cPane) {
-		this.setSize(800, 800);
-		this.setDefaultCloseOperation(JFrame.DO_NOTHING_ON_CLOSE);
-		this.addWindowListener(new WindowListener() {
-			public void windowOpened(WindowEvent e) {}
-			public void windowClosing(WindowEvent e) {
-				int closeOpt = JOptionPane.showOptionDialog(cPane, "Would you like to Save, Exit, or Cancel", 
-						"Exit Confirmation", JOptionPane.DEFAULT_OPTION, JOptionPane.QUESTION_MESSAGE, 
-						null, new String[] {"Save", "Exit Without Saving", "Cancel"}, 3);
-				if(closeOpt == 0) {
-					WriteMonsters();
-					data.shutDownAndWait();
-					dispose();
-					System.exit(0);
-				}else if(closeOpt == 1) {
-					dispose();
-					data.shutDownAndWait();
-					System.exit(0);
-				}
-			}
-			public void windowClosed(WindowEvent e) {}
-			public void windowIconified(WindowEvent e) {}
-			public void windowDeiconified(WindowEvent e) {}
-			public void windowActivated(WindowEvent e) {}
-			public void windowDeactivated(WindowEvent e) {}			
-		});
-		cPane.setLayout(new BorderLayout());
+	private void ConfigurePane() {
+
+		setLayout(new BorderLayout());
 
 		tabPane = new JTabbedPane();
 //		StyleContainer.SetLookAndFeel();
@@ -167,11 +141,11 @@ public class MonsterBuilder extends JFrame {
 		BuildTagsPane(tagPane);
 		tabPane.addTab("Tags", tagPane);
 
-		cPane.add(tabPane, BorderLayout.CENTER);
+		add(tabPane, BorderLayout.CENTER);
 
 		JPanel addBtnPane = new JPanel();
 		addBtnPane.setLayout(new FlowLayout(FlowLayout.RIGHT));
-		cPane.add(addBtnPane, BorderLayout.SOUTH);
+		add(addBtnPane, BorderLayout.SOUTH);
 		
 		JButton addMonster = new JButton("Add Monster");
 		StyleContainer.SetFontBtn(addMonster);
@@ -203,7 +177,7 @@ public class MonsterBuilder extends JFrame {
 		sidePane.setLayout(new BorderLayout());
 		monstGridPane = new JPanel();
 		LoadMonsters(data.getMonsters());
-		cPane.add(sidePane, BorderLayout.WEST);
+		add(sidePane, BorderLayout.WEST);
 		
 		JScrollPane monstScroll = CompFactory.wrapPanelInScroll(monstGridPane, ScrollPolicy.VERTICAL);
 		sidePane.add(monstScroll, BorderLayout.CENTER);
@@ -399,6 +373,11 @@ public class MonsterBuilder extends JFrame {
 
 		JPanel tPane = new JPanel();
 		tPane.setLayout(new BorderLayout());
+		
+		JPanel headerPane = new JPanel();
+		headerPane.setLayout(new BorderLayout());
+		tPane.add(headerPane, BorderLayout.CENTER);
+		
 		monsterNameField = new ReminderField("", "Enter the name for the monster...");
 		StyleContainer.SetFontHeader(monsterNameField);
 		monsterNameField.setFont(monsterNameField.getFont().deriveFont((float) 22));
@@ -410,7 +389,16 @@ public class MonsterBuilder extends JFrame {
 				rEditReps.put("<NAME>", monsterNameField.getText().toLowerCase());
 			}
 		});
-		tPane.add(monsterNameField, BorderLayout.CENTER);
+		headerPane.add(monsterNameField, BorderLayout.CENTER);
+		
+		JPanel srcPane = new JPanel();
+		srcPane.setLayout(new BorderLayout());
+		headerPane.add(srcPane, BorderLayout.EAST);
+		
+		srcPane.add(CompFactory.createNewLabel("Source:", ComponentType.BODY), BorderLayout.WEST);
+		
+		srcCombo = CompFactory.createEnumCombo(Source.class, ComponentType.BODY);
+		srcPane.add(srcCombo, BorderLayout.CENTER);
 
 		monsterTypeField = new ReminderField("", "Enter the monster size type, alignment");
 		monsterTypeField.setFont(StyleContainer.FNT_BODY_PLAIN.deriveFont(Font.ITALIC));
@@ -818,39 +806,31 @@ public class MonsterBuilder extends JFrame {
 	private void BuildMonstListPane() {
 		SwingUtilities.invokeLater(()->{
 			monstGridPane.removeAll();
-			monstGridPane.setLayout(new GridLayout(0, 2));
+			monstGridPane.setLayout(new GridLayout(0, 1));
 			
-			ArrayList<String> sortKeys = new ArrayList<String>();
-			for(String key : monstMap.keySet()) {
-				sortKeys.add(key);
-			}
+			ArrayList<String> sortKeys = new ArrayList<String>(monstMap.keySet());
 			Collections.sort(sortKeys);
+			
 			for (String key : sortKeys) {
+				JPanel pane = new JPanel();
+				pane.setLayout(new BorderLayout());
+				monstGridPane.add(pane);
+				
 //				System.out.println(key);
-				JTextField monstName = new JTextField(key);
-				StyleContainer.SetFontMain(monstName);
-				monstName.setColumns(15);
-				monstName.addMouseListener(new MouseListener() {
+				String name = key;
+				if(name.length() > 13)
+					name = name.substring(0, 13);
+				
+				JLabel monstName = CompFactory.createSideLabel(name, ComponentType.BODY);
+				monstName.addMouseListener(CompFactory.createSideMouseListener(monstName, ()->{
+					int opt = JOptionPane.showConfirmDialog(null, 
+								"Load: " + key + "? Any unadded work on current monster will be lost.", 
+								"Load Confirm", JOptionPane.YES_NO_OPTION);
+						if(opt == JOptionPane.YES_OPTION)
+							LoadEditMonster(monstMap.get(key));
+				}));
 
-					@Override
-					public void mouseClicked(MouseEvent e) {
-						JFrame monstDisp = new JFrame();
-						monstDisp.setTitle(key);
-						monstDisp.setContentPane(new MonsterDispPane(monstMap.get(key), data, new GuiDirector(new JDesktopPane())));
-						monstDisp.setSize(645, 515);
-						monstDisp.setDefaultCloseOperation(JFrame.DISPOSE_ON_CLOSE);						
-						monstDisp.setVisible(true);
-						monstDisp.setResizable(false);
-					}
-					public void mousePressed(MouseEvent e) {}
-					public void mouseReleased(MouseEvent e) {}
-					public void mouseEntered(MouseEvent e) {monstName.setFont(monstName.getFont().deriveFont(Font.BOLD));}
-					public void mouseExited(MouseEvent e) {monstName.setFont(monstName.getFont().deriveFont(Font.PLAIN));}
-				});
-				monstName.setEditable(false);
-				monstName.setFocusable(false);
-
-				monstGridPane.add(monstName);
+				pane.add(monstName, BorderLayout.CENTER);
 
 				JButton mDel = new JButton("Delete");
 				StyleContainer.SetFontBtn(mDel);
@@ -869,7 +849,7 @@ public class MonsterBuilder extends JFrame {
 					}
 					
 				});
-				monstGridPane.add(mDel);
+				pane.add(mDel, BorderLayout.EAST);
 			}
 			monstGridPane.repaint();
 			monstGridPane.revalidate();
@@ -938,6 +918,7 @@ public class MonsterBuilder extends JFrame {
 			mData.type = monsterTypeField.getText();
 			mData.wis = Integer.parseInt(wisField.getText());
 			mData.wisSProf = wisSaveP.isSelected();
+			mData.source = (Source) srcCombo.getSelectedItem();
 
 			Monster m = MonsterFactory.BuildMonster(mData);
 			System.out.println("Return true, add monst");
@@ -989,28 +970,6 @@ public class MonsterBuilder extends JFrame {
 	}
 	
 	public boolean WriteMonsters() {
-//		File monstFile = new File(DataContainer.MONSTERS_FILE_NAME);
-//		if(!monstFile.exists()) {
-//			try {
-//				monstFile.createNewFile();
-//			} catch (IOException e) {
-//				e.printStackTrace();
-//			}
-//		}
-//		
-//		try {
-//			ObjectOutputStream oos = new ObjectOutputStream(new FileOutputStream(monstFile));
-//			for (String s : monstMap.keySet()) {
-//				oos.writeObject(monstMap.get(s));
-//			}
-//			oos.flush();
-//			oos.close();
-//			return true;
-//		} catch (IOException e) {
-//			e.printStackTrace();
-//			return false;
-//		}
-		
 		data.SetMonstersMap(monstMap);
 		data.SafeSaveData(MapType.MONSTERS);
 		return true;
@@ -1079,6 +1038,116 @@ public class MonsterBuilder extends JFrame {
 		
 		repaint();
 		revalidate();
+	}
+	
+	private void LoadEditMonster(Monster m) {
+		ResetEditor();
+		monsterNameField.setText(m.name);
+		monsterNameField.setEditable(false);
+		monsterNameField.setFocusable(false);
+		monsterTypeField.setText(m.typeSizeAlignment);
+		srcCombo.setSelectedItem(m.source);
+		System.out.println(m.source.toString());
+		acField.setText(m.ac);
+		initBnsField.setText("0");
+		hpField.setText(m.hp);
+		speedField.setText(m.speed);
+		strField.setText("" + m.stats[Monster.STR]);
+		dexField.setText("" + m.stats[Monster.DEX]);
+		conField.setText("" + m.stats[Monster.CON]);
+		intField.setText("" + m.stats[Monster.INT]);
+		wisField.setText("" + m.stats[Monster.WIS]);
+		chaField.setText("" + m.stats[Monster.CHA]);
+		dmgResistField.setText(m.dmgRes);
+		dmgVulnField.setText(m.dmgVul);
+		immuneField.setText(m.immune);
+		sensesField.setText(m.senses);
+		langField.setText(m.languages);
+		crField.setText(m.cr + "");
+		numLActs.setText(m.lActNum + "");
+		numLActsBns.setText(m.lActBns + "");
+		int prof = m.profByCR();
+		int init = m.GetInitBonus();
+		int dex = m.stats[Monster.DEX];
+		if(init > Monster.AbilityModCalc(dex)) {
+			if(prof+Monster.AbilityModCalc(dex) == init) {
+				initProfChck.setSelected(true);
+				initExpertChck.setSelected(false);
+			}else {
+				initProfChck.setSelected(false);
+				initExpertChck.setSelected(true);
+			}
+		}
+		strSaveP.setSelected(m.saves[Monster.STR]);
+		dexSaveP.setSelected(m.saves[Monster.DEX]);
+		conSaveP.setSelected(m.saves[Monster.CON]);
+		intSaveP.setSelected(m.saves[Monster.INT]);
+		wisSaveP.setSelected(m.saves[Monster.WIS]);
+		chaSaveP.setSelected(m.saves[Monster.CHA]);
+		
+//		traitsPane.remove(traits);
+		removeAndDergister(traitsPane, traits);
+		traits = new RichEditor(data, rEditReps);
+		traits.LoadDocument(m.traits);
+		traitsPane.add(traits);
+//		actPane.remove(actions);
+		removeAndDergister(actPane, actions);
+		actions = new RichEditor(data, rEditReps);
+		actions.LoadDocument(m.actions);
+		actPane.add(actions);
+//		baPane.remove(bonusActions);
+		removeAndDergister(baPane, bonusActions);
+		bonusActions = new RichEditor(data, rEditReps);
+		bonusActions.LoadDocument(m.bonusActions);
+		baPane.add(bonusActions);
+//		reactPane.remove(reactions);
+		removeAndDergister(reactPane, reactions);
+		reactions = new RichEditor(data, rEditReps);
+		reactions.LoadDocument(m.reactions);
+		reactPane.add(reactions);
+//		legendActPane.remove(legendActions);
+		removeAndDergister(legendActPane, legendActions);
+		legendActions = new RichEditor(data, rEditReps);
+		legendActions.LoadDocument(m.legendActions);
+		legendActPane.add(legendActions);
+		
+		srcCombo.setSelectedItem(m.source);
+		
+		skills = m.skills;
+		tagFields.clear();
+		
+		for(String s : m.tags)
+			AddTag(s);
+		
+		tabPane.setSelectedIndex(0);
+		
+		repaint();
+		revalidate();
+	}
+	
+	private void AddTag(String s) {
+		ReminderField tagField = new ReminderField("", "Enter Tag");
+		tagField.setText(s);
+		StyleContainer.SetFontHeader(tagField);
+
+		JButton dTagBtn = new JButton("Delete");
+		StyleContainer.SetFontBtn(dTagBtn);
+		dTagBtn.setFocusable(false);
+		dTagBtn.addActionListener(e2 -> {
+			SwingUtilities.invokeLater(() -> {
+				tagFields.remove(tagField);
+				tagGrid.remove(tagField);
+				tagGrid.remove(dTagBtn);
+
+				tagGrid.revalidate();
+				tagGrid.repaint();
+			});
+		});
+
+		tagGrid.add(tagField);
+		tagGrid.add(dTagBtn);
+		tagField.requestFocus();
+		tagFields.add(tagField);
 	}
 	
 	public void removeAndDergister(JPanel pane, RichEditor r) {
